requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
